feat(scout): charts adapter (openswap #16) + close the audit's failing-test hole
TOOLING FIX (the important half): goat_audit.py counted assertions but never RAN
them, so a plugin whose suite was red could score a perfect 10.00. Combined with
ci.yml's pytest step ending in `|| true`, a broken plugin passed on BOTH paths —
false assurance sitting inside the very tool meant to catch it. Caught live: the
audit reported charts 10.00 while 3 of its tests were failing.

New --run-tests executes each plugin's suite and forces D4 to 0 on failure, with
the pytest tail in the finding. Opt-in because each suite runs ~45-90s; the CI
gate scopes it with --plugin. Verified end-to-end against charts in both states.

charts (openswap #16, replaces Grafana Cloud / Tableau): line/bar/scatter emitted
as string-template SVG, no chart library, deterministic output so charts are
diffable, provenance footer naming the source and row count on every chart.

VERIFIED, not self-reported: 57 tests pass; goat_audit --plugin charts --run-tests
= 10.00 with suite_passed=true; stdlib-only (grepped for numpy/matplotlib/pandas/
requests — none). The builder had a brittle assertion on exact SVG byte length
(3056 == 3062) which it fixed to be robust rather than pinned.

Batch 3 remainder NOT committed, honestly: alerts 8.33, headers 8.33, ollama 8.17,
metrics 8.17 — all four still have NO tests, so none clears the >=9.0 gate.
readability has no plugin dir, which may be correct if it chose to extend prose.

+++ apps/scout-cli/bigbang/core/charts.py
@@ -642,7 +642,7 @@ def nice_axis(lo: float, hi: float, *, count: int = DEFAULT_TICKS) -> dict[str, 
     step = _nice_step((hi - lo) / count)
     nlo = math.floor(lo / step) * step
     nhi = math.ceil(hi / step) * step
-    n = int(round((nhi - nlo) / step))
+    n = round((nhi - nlo) / step)
     n = max(1, min(n, MAX_TICKS))
     ticks = [round(nlo + i * step, 12) for i in range(n + 1)]
     return {"lo": round(nlo, 12), "hi": round(ticks[-1], 12), "step": step,
